Add utils and rename files
